refactor(tldraw): remove the TypeScript freehand algorithm; WASM is the sole implementation
The freehand generators (getStroke, getStrokePoints, getStrokeOutlinePoints,
getSvgPathFromStrokePoints, getSvgPathFromPoints, and the draw/highlight render
paths) are now thin wrappers over the Rust/WASM module with no JS fallback. The
TypeScript algorithm (core, fmt, svgInk, and the outline builder) is removed
from the shipped SDK.

The JS implementation is kept verbatim under freehand/oracle/ purely as a test
oracle: the parity tests compare the WASM output against it byte-for-byte (and
within 1e-6 for point arrays). It is not exported and not reachable from the
package entry point, so it isn't part of the shipped bundle or public API.
