+++ packages/editor/src/lib/utils/getSvgPathFromPoints.ts
@@ -1,4 +1,3 @@
-import { average, precise } from '../primitives/utils'
 import { VecLike } from '../primitives/Vec'
 import { getSvgPathFromPointsWasm } from './freehand-wasm/svgInkWasm'
 
@@ -11,45 +10,6 @@ import { getSvgPathFromPointsWasm } from './freehand-wasm/svgInkWasm'
  * @public
  */
 export function getSvgPathFromPoints(points: VecLike[], closed = true): string {
-	// The path generation is implemented in Rust/WASM (see freehand-wasm); this JS body is a
-	// fallback for environments that can't instantiate the module.
-	const wasm = getSvgPathFromPointsWasm(points, closed)
-	if (wasm !== null) return wasm
-
-	const len = points.length
-
-	if (len < 2) {
-		return ''
-	}
-
-	let a = points[0]
-	let b = points[1]
-
-	if (len === 2) {
-		// If only two points, just draw a line
-		return `M${precise(a)}L${precise(b)}`
-	}
-
-	let result = ''
-
-	for (let i = 2, max = len - 1; i < max; i++) {
-		a = points[i]
-		b = points[i + 1]
-		result += average(a, b)
-	}
-
-	if (closed) {
-		// If closed, draw a curve from the last point to the first
-		return `M${average(points[0], points[1])}Q${precise(points[1])}${average(
-			points[1],
-			points[2]
-		)}T${result}${average(points[len - 1], points[0])}${average(points[0], points[1])}Z`
-	} else {
-		// If not closed, draw a curve starting at the first point and
-		// ending at the midpoint of the last and second-last point, then
-		// complete the curve with a line segment to the last point.
-		return `M${precise(points[0])}Q${precise(points[1])}${average(points[1], points[2])}${
-			points.length > 3 ? 'T' : ''
-		}${result}L${precise(points[len - 1])}`
-	}
+	// Path generation is implemented in Rust/WASM (see freehand-wasm).
+	return getSvgPathFromPointsWasm(points, closed) ?? ''
 }
