+++ packages/tldraw/src/lib/shapes/shared/freehand/getStroke.ts
@@ -1,6 +1,4 @@
 import { Vec, VecLike, strokeOutlineWasm } from '@tldraw/editor'
-import { computeRadii, ingest, loadSrcFromPipeline, pointCount } from './core'
-import { outlineFromSrc } from './getStrokeOutlinePoints'
 import type { StrokeOptions } from './types'
 
 /**
@@ -13,20 +11,10 @@ import type { StrokeOptions } from './types'
  * @param options - An object with options.
  * @public
  */
-
 export function getStroke(points: VecLike[], options: StrokeOptions = {}): Vec[] {
-	const data = strokeOutlineWasm(points, options)
-	if (data) {
-		const n = data.length / 2
-		const out: Vec[] = new Array(n)
-		for (let i = 0; i < n; i++) out[i] = new Vec(data[i * 2], data[i * 2 + 1])
-		return out
-	}
-
-	// Fallback for environments that can't instantiate the WASM module.
-	ingest(points, options)
-	if (pointCount === 0) return []
-	computeRadii(options)
-	loadSrcFromPipeline()
-	return outlineFromSrc(options)
+	const data = strokeOutlineWasm(points, options) ?? new Float64Array(0)
+	const n = data.length / 2
+	const out: Vec[] = new Array(n)
+	for (let i = 0; i < n; i++) out[i] = new Vec(data[i * 2], data[i * 2 + 1])
+	return out
 }
