+++ packages/tldraw/src/lib/shapes/shared/freehand/oracle/fmt.ts
